# backend/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package db

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type FeatureFlag struct {
	Key         string    `json:"key"`
	Enabled     bool      `json:"enabled"`
	Description string    `json:"description"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}

type RefreshToken struct {
	ID        uuid.UUID          `json:"id"`
	UserID    uuid.UUID          `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt time.Time          `json:"expires_at"`
	Revoked   pgtype.Bool        `json:"revoked"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID                        uuid.UUID          `json:"id"`
	Email                     string             `json:"email"`
	PasswordHash              string             `json:"password_hash"`
	Type                      interface{}        `json:"type"`
	EmailVerified             pgtype.Bool        `json:"email_verified"`
	FirstName                 pgtype.Text        `json:"first_name"`
	LastName                  pgtype.Text        `json:"last_name"`
	AvatarUrl                 pgtype.Text        `json:"avatar_url"`
	CreatedAt                 pgtype.Timestamptz `json:"created_at"`
	UpdatedAt                 pgtype.Timestamptz `json:"updated_at"`
	PasswordResetSelector     pgtype.Text        `json:"password_reset_selector"`
	PasswordResetVerifierHash pgtype.Text        `json:"password_reset_verifier_hash"`
	PasswordResetExpires      pgtype.Timestamptz `json:"password_reset_expires"`
	VerificationSelector      pgtype.Text        `json:"verification_selector"`
	VerificationVerifierHash  pgtype.Text        `json:"verification_verifier_hash"`
}